requirements.txt
streamlit
openai
python-dotenv
requests
pydantic
